Validate treasury executor envs in deploy readiness
